requirements.txt
mcstatus~=11.0.1
discord~=2.3.2
python-dotenv~=1.0.0